Exercício 2 - Chat Bot/requirements.txt
langchain==0.3.0
langchain-groq==0.2.0
python-dotenv
